requirements.txt
pyzmq
gpiozero